requirements.txt
argon2-cffi>=25.1.0
asyncpg>=0.30.0
cryptography>=45.0.3
dotenv>=0.9.9
fastapi>=0.115.12
fastapi_mail>=1.5.0
httpx>=0.28.1
pydantic[email]
pyotp>=2.9.0
sqlalchemy>=2.0.41
uvicorn>=0.34.2
